+++ src/main/java/stsjorbsmod/cards/cull/deckoftrials/DeckOfTrials.java
@@ -16,6 +16,7 @@ public class DeckOfTrials {
         deck.addToTop(new Investment());
         deck.addToTop(new ProdigalMemory());
         deck.addToTop(new TickingCurse());
+        deck.addToTop(new ShrapnelBloom());
         deck.addToTop(new Assertion());
         deck.shuffle();
     }